pyproject.toml
[project]
name = "tube-song-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "azure-ai-textanalytics>=5.3.0"
]

[tool.setuptools]
py-modules = ["main", "youtube_comments", "sentiment_analyzer"]
